Изменен метод добавления пользователей, избавился от лишних переменных
